Fix bug upload content-type bug.
The content-type was being detected during the first iteration of
the loop, then remaining as the content-type that was first
detected.

This resulted from re-using global upload options set outside of
the upload loop. The fix involved creating a new set of upload
options within the loop and utilizing global upload settings
separately, copying them into the new options object as needed.